Refacring logout

+++ src/services/local-storage.service.ts
@@ -1,9 +1,15 @@
+import { LS_APP_USER } from "./../common/constants/localStorage";
 export const setKey = (key: string, value: string) => {
     localStorage.setItem(key, value);
 };
 
-export const clear = () => {
-    localStorage.clear();
+type LocalStorageKeys = typeof LS_APP_USER;
+
+export const localStorageClear = (key?: LocalStorageKeys) => {
+    if (!key) {
+        return localStorage.clear();
+    }
+    localStorage.removeItem(key);
 };
 
 export const getKeyValue = (key: string) => {
